Support 2018 tables for institutions-api in beta namespace
